backend/requirements.txt
Flask==3.0.0
google-generativeai==0.8.3
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
requests==2.31.0
